// src/lib/enums/osuTypes.ts
enum osuTypes {
    i8 = 0,
    u8 = 1,
    i16 = 2,
    u16 = 3,
    i32 = 4,
    u32 = 5,
    f32 = 6,
    i64 = 7,
    u64 = 8,
    f64 = 9,

    message = 11,
    channel = 12,
    match = 13,
    scoreframe = 14,
    mapInfoRequest = 15,
    mapInfoReply = 16,
    replayFrameBundle = 17,

    i32Array = 18,
    i32Array4L = 19,
    string = 20,
    raw = 21
}

export default osuTypes;